created a class to resolve a jvm bug that could rarely show up during deserialization and refactored serialization and deserialization classes
