+++ src/test/scala/controller/util/serialization/StoryNodeSerializerTest.scala
@@ -1,7 +1,7 @@
-package model.nodes.util
+package controller.util.serialization
 
 import controller.util.ResourceName
-import model.nodes.util.StoryNodeSerializer.{deserializeStory, serializeStory}
+import controller.util.serialization.StoryNodeSerializer.{deserializeStory, serializeStory}
 import model.nodes.{Pathway, StoryNode}
 import specs.FlatTestSpec
 import util.TestUtil.deleteFolder
@@ -11,7 +11,6 @@ import java.io.{File, FileNotFoundException}
 class StoryNodeSerializerTest extends FlatTestSpec {
   private val serializationFileNameTest =  "serializationTest"
   private val serializationPathTest: String = ResourceName.TempDirectory + "/" + serializationFileNameTest
-
   private val serializationFile = new File(serializationPathTest)
 
   val destinationNodeCondition: StoryNode = StoryNode(2, "narrative", Set.empty)
